+++ spec/emarsys/data_objects/export_spec.rb
@@ -16,4 +16,16 @@ describe Emarsys::Export do
       ).to have_been_requested.once
     end
   end
+
+  describe ".filter" do
+    it "requests a single export" do
+      stub = stub_request(:post, "https://api.emarsys.net/api/v2/export/filter").to_return(standard_return_body)
+      Emarsys::Export.filter(
+        filter: 123,
+        distribution_method: 'local',
+        contact_fields: [1, 3]
+      )
+      expect(stub).to have_been_requested.once
+    end
+  end
 end